move TR and OL special cases into didSplit rules
