+++ src/flow-box/clonePath.js
@@ -1,31 +1,5 @@
 import { classes } from '../dom-utils';
 
-const preserveNumbering = (original, clone, nextChild) => {
-  // restart numbering
-  let prevStart = 1;
-  if (original.hasAttribute('start')) {
-    // the OL is also a continuation
-    prevStart = parseInt(original.getAttribute('start'), 10);
-  }
-  if (nextChild && nextChild.tagName === 'LI') {
-    // the first list item is a continuation
-    prevStart -= 1;
-  }
-  const prevCount = original.children.length;
-  const newStart = prevStart + prevCount;
-  clone.setAttribute('start', newStart);
-};
-
-const preserveTableColumns = (original, clone, nextChild, deepClone) => {
-  const columns = [...original.children];
-  const currentIndex = columns.indexOf(nextChild);
-  for (let i = 0; i < currentIndex; i += 1) {
-    const clonedCol = deepClone(columns[i]);
-    clone.appendChild(clonedCol);
-  }
-};
-
-
 // The path is an array of nested elments,
 // for example .content > article > p > a).
 //
@@ -51,10 +25,10 @@ const clonePath = (oldPath, applyRules) => {
     // fromPrev.forEach(cl => node.classList.add(cl));
   };
 
-  const finishSplitting = (original, clone) => {
+  const finishSplitting = (original, clone, nextChild, cloneEl) => {
     markAsToNext(original);
     markAsFromPrev(clone);
-    applyRules(original, clone);
+    applyRules(original, clone, nextChild, cloneEl);
   };
 
   const deepClone = (el) => {
@@ -66,19 +40,10 @@ const clonePath = (oldPath, applyRules) => {
   for (let i = oldPath.length - 1; i >= 0; i -= 1) {
     const original = oldPath[i];
     const clone = original.cloneNode(false); // shallow
+    const nextChild = oldPath[i + 1];
     clone.innerHTML = '';
 
-    finishSplitting(original, clone);
-
-    // Special case for ordered lists
-    if (clone.tagName === 'OL') {
-      preserveNumbering(original, clone, oldPath[i + 1]);
-    }
-
-    // Special case to preserve columns of tables
-    if (clone.tagName === 'TR') {
-      preserveTableColumns(original, clone, oldPath[i + 1], deepClone);
-    }
+    finishSplitting(original, clone, nextChild, deepClone);
 
     if (i < oldPath.length - 1) clone.appendChild(newPath[i + 1]);
     newPath[i] = clone;
